Created several base functions. Added food and thorn factory objects.

+++ AgarIoClone/src/server/model/helper/Coords.java
@@ -3,9 +3,9 @@ package server.model.helper;
 
 public class Coords {
     
-    private int x,y;
+    private float x,y;
     
-    public Coords(int x, int y){
+    public Coords(float x, float y){
         this.x=x;
         this.y=y;
     }
@@ -18,11 +18,11 @@ public class Coords {
         this.y=y;
     }
     
-    public int getX(){
+    public float getX(){
         return this.x;
     }
     
-    public int getY(){
+    public float getY(){
         return this.y;
     }
